Change monetary properties to BigDecimal

+++ src/main/java/com/mercadopago/resources/payment/PaymentTax.java
@@ -1,5 +1,6 @@
 package com.mercadopago.resources.payment;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 /** PaymentTax class. */
@@ -7,5 +8,5 @@ import lombok.Data;
 public class PaymentTax {
   private String type;
 
-  private Double value;
+  private BigDecimal value;
 }
